requirements.txt
streamlit
langchain
openai
beautifulsoup4
requests
python-dotenv
pinecone 